% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ethirgast.R
\name{run_ethirgast}
\alias{run_ethirgast}
\title{Run the ETHIRGAST Shiny Application}
\usage{
run_ethirgast()
}
\value{
An object representing the ETHIRGAST app
}
\description{
Run the ETHIRGAST Shiny Application
}
