% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startpage.R
\name{startpage_server}
\alias{startpage_server}
\title{Server-part av startpage for ETHIRGAST's Shiny-application}
\usage{
startpage_server(id)
}
\description{
Server-part av startpage for ETHIRGAST's Shiny-application
}
